Extraction des délivrances de médicaments.

Cette fonction permet d'extraire les délivrances de médicaments. Les délivrances dont les dates EXE_SOI_DTD sont comprises entre start_date et end_date (incluses) sont extraites. Le décalage de remontée des données est pris en compte en récupérant également les délivrances dont les dates FLX_DIS_DTD sont comprises dans les dis_dtd_lag_months mois suivant end_date. Si atc_cod_starts_with est fourni, seules les délivrances de médicaments dont le code ATC commence par l'un des éléments de atc_cod_starts_with sont extraites. Dans le cas contraire, les délivrances pour tous les codes ATC sont extraites. Si patients_ids est fourni, seules les délivrances de médicaments pour les patients dont les identifiants sont dans patients_ids sont extraites. Dans le cas contraire, les délivrances de tous les patiens sont extraites.

Utilisation

extract_drug_dispenses(
  start_date = NULL,
  end_date = NULL,
  atc_cod_starts_with = NULL,
  dis_dtd_lag_months = 7,
  patients_ids = NULL,
  output_table_name = NULL,
  conn = NULL
)

Arguments

start_date

Date La date de début de la période des délivrances des médicaments à extraire.

end_date

Date La date de fin de la période des délivrances des médicaments à extraire.

atc_cod_starts_with

Character vector Optionnel. Les codes ATC par lesquels les délivrances de médicaments à extraire doivent commencer.

dis_dtd_lag_months

Integer Optionnel. Le nombre maximum de mois de décalage de FLX_DIS_DTD par rapport à EXE_SOI DTD pris en compte pour récupérer les délivrances de médicaments. Par défaut, 7 mois.

patients_ids

data.frame Optionnel. Un data.frame contenant les paires d'identifiants des patients pour lesquels les délivrances de médicaments doivent être extraites. Les colonnes de ce data.frame doivent être "ben_idt_ano" et "ben_nir_psa" (en minuscules). Les "ben_nir_psa" doivent être tous les "ben_nir_psa" associés aux "ben_idt_ano" fournis.

output_table_name

Character Optionnel. Si fourni, les résultats seront sauvegardés dans une table portant ce nom dans la base de données au lieu d'être retournés sous forme de data frame.

conn

DBI connection Une connexion à la base de données Oracle. Si non fournie, une connexion est établie par défaut.

Valeur de retour

Si output_table_name est NULL, retourne un data.frame contenant les délivrances de médicaments. Si output_table_name est fourni, sauvegarde les résultats dans la table spécifiée dans Oracle et retourne NULL de manière invisible. Dans les deux cas les colonnes de la table de sortie sont :

BEN_NIR_PSA : Colonne présente uniquement si les identifiants patients (patients_ids) ne sont pas fournis. Identifiant SNDS, ausi appelé pseudo-NIR.

BEN_IDT_ANO : Colonne présente uniquement si les identifiants patients (patients_ids) sont fournis. Numéro d’inscription au répertoire (NIR) anonymisé.

EXE_SOI_DTD : Date de la délivrance

PHA_ACT_QSN : Quantité délivrée

PHA_ATC_CLA : Code ATC du médicament délivré

PHA_CIP_C13 : Code CIP du médicament délivré (nom dans la table IR_PHA_R : PHA_CIP_C13, nom dans la table ER_PRS_F : PHA_PRS_C13)

PSP_SPE_COD : Code de spécialité du professionnel de soin prescripteur (voir nomenclature IR_SPE_V)

Exemples

if (FALSE) { # \dontrun{
start_date <- as.Date("2010-01-01")
end_date <- as.Date("2010-01-03")
atc_cod_starts_with <- c("N04A")

dispenses <- extract_drug_dispenses(
  start_date = start_date,
  end_date = end_date,
  atc_cod_starts_with = atc_cod_starts_with
)
} # }